Separate the concurrency-compatible state converter from the state manager
