+++ airbyte-cdk/python/unit_tests/sources/streams/concurrent/scenarios/incremental_scenarios.py
@@ -32,18 +32,10 @@ test_incremental_stream_without_slice_boundaries_no_input_state = (
                 )
             ]
         )
-        .set_incremental(CursorField(["cursor_field"]), _NO_SLICE_BOUNDARIES)
+        .set_incremental(CursorField("cursor_field"), _NO_SLICE_BOUNDARIES)
         .set_input_state(_NO_INPUT_STATE)
     )
-    .set_expected_records(
-        [
-            {"data": {"id": "1", "cursor_field": 0}, "stream": "stream1"},
-            {"data": {"id": "2", "cursor_field": 1}, "stream": "stream1"},
-            {"data": {"id": "3", "cursor_field": 2}, "stream": "stream1"},
-            {"data": {"id": "4", "cursor_field": 3}, "stream": "stream1"},
-            {"stream1": {"slices": [{"start": 0, "end": 3}], "state_type": ConcurrencyCompatibleStateType.date_range.value, "legacy": {}}},
-        ]
-    )
+    .set_expected_read_error(ValueError, "test exception")
     .set_log_levels({"ERROR", "WARN", "WARNING", "INFO", "DEBUG"})
     .set_incremental_scenario_config(IncrementalScenarioConfig(input_state=_NO_INPUT_STATE))
     .build()
@@ -73,7 +65,7 @@ test_incremental_stream_with_slice_boundaries_no_input_state = (
                 )
             ]
         )
-        .set_incremental(CursorField(["cursor_field"]), ("from", "to"))
+        .set_incremental(CursorField("cursor_field"), ("from", "to"))
         .set_input_state(_NO_INPUT_STATE)
     )
     .set_expected_records(
@@ -116,24 +108,10 @@ test_incremental_stream_without_slice_boundaries_with_legacy_state = (
                 )
             ]
         )
-        .set_incremental(CursorField(["cursor_field"]), _NO_SLICE_BOUNDARIES)
+        .set_incremental(CursorField("cursor_field"), _NO_SLICE_BOUNDARIES)
         .set_input_state(LEGACY_STATE)
     )
-    .set_expected_records(
-        [
-            {"data": {"id": "1", "cursor_field": 0}, "stream": "stream1"},
-            {"data": {"id": "2", "cursor_field": 1}, "stream": "stream1"},
-            {"data": {"id": "3", "cursor_field": 2}, "stream": "stream1"},
-            {"data": {"id": "4", "cursor_field": 3}, "stream": "stream1"},
-            {
-                "stream1": {
-                    "slices": [{"start": 0, "end": 3}],
-                    "state_type": ConcurrencyCompatibleStateType.date_range.value,
-                    "legacy": {"created": 0},
-                }
-            },
-        ]
-    )
+    .set_expected_read_error(ValueError, "test exception")
     .set_log_levels({"ERROR", "WARN", "WARNING", "INFO", "DEBUG"})
     .set_incremental_scenario_config(IncrementalScenarioConfig(input_state=LEGACY_STATE))
     .build()
@@ -229,18 +207,10 @@ test_incremental_stream_without_slice_boundaries_with_concurrent_state = (
                 )
             ]
         )
-        .set_incremental(CursorField(["cursor_field"]), _NO_SLICE_BOUNDARIES)
+        .set_incremental(CursorField("cursor_field"), _NO_SLICE_BOUNDARIES)
         .set_input_state(CONCURRENT_STATE)
     )
-    .set_expected_records(
-        [
-            {"data": {"id": "1", "cursor_field": 0}, "stream": "stream1"},
-            {"data": {"id": "2", "cursor_field": 1}, "stream": "stream1"},
-            {"data": {"id": "3", "cursor_field": 2}, "stream": "stream1"},
-            {"data": {"id": "4", "cursor_field": 3}, "stream": "stream1"},
-            {"stream1": {"slices": [{"start": 0, "end": 3}], "state_type": ConcurrencyCompatibleStateType.date_range.value}},
-        ]
-    )
+    .set_expected_read_error(ValueError, "test exception")
     .set_log_levels({"ERROR", "WARN", "WARNING", "INFO", "DEBUG"})
     .set_incremental_scenario_config(IncrementalScenarioConfig(input_state=CONCURRENT_STATE))
     .build()
